Add new client dependencies with build-android-deps Refactoring
